+++ scripts/datacite-doi.py
@@ -5,7 +5,7 @@ import json
 import logging
 from pathlib import Path
 import datacite.config
-from datacite.doi import Doi
+from datacite.doi import State, Doi
 import datacite.xml
 
 logging.basicConfig(level=logging.DEBUG, format="%(levelname)s: %(message)s")
@@ -17,7 +17,7 @@ subparsers = argparser.add_subparsers(title='subcommands', dest='subcmd')
 
 
 def get_metadata(args):
-    config = datacite.config.get_config(args, login=False)
+    config = datacite.config.get_config(args, login=args.login)
     doi = Doi(args.doi)
     doi.fetch(config)
     if args.show == 'overview':
@@ -33,6 +33,9 @@ def get_metadata(args):
                            % args.show)
 
 get_metadata_parser = subparsers.add_parser('get', help="Query a DOI")
+get_metadata_parser.add_argument('--login', action='store_true',
+                                 help=("login to DataCite in order to "
+                                       "search for the DOI"))
 get_metadata_parser.add_argument('--show',
                                  help="select the kind of information to show",
                                  choices=['overview',
@@ -45,20 +48,27 @@ get_metadata_parser.set_defaults(func=get_metadata)
 def create_doi(args):
     config = datacite.config.get_config(args)
     doi = Doi(args.doi)
-    doi.url = args.url
-    metadata = datacite.xml.XML(args.metadata)
-    metadata.doi = doi.doi
-    doi.metadata = metadata
-    log.info("Mint %s for %s", doi.doi, metadata.title)
-    doi.create(config)
+    if args.url:
+        doi.url = args.url
+    if args.metadata:
+        metadata = datacite.xml.XML(args.metadata)
+        metadata.doi = doi.doi
+        doi.metadata = metadata
+    log.info("Mint %s for %s",
+             doi.doi, (metadata.title if args.metadata else "N/A"))
+    doi.create(config, state=args.state)
 
 create_parser = subparsers.add_parser('create', help="Mint a DOI")
-create_parser.add_argument('doi', help="the DOI to create")
-create_parser.add_argument('url', help="URL of the landing page")
-create_parser.add_argument('metadata',
+create_parser.add_argument('--state',
+                           type=State, choices=[str(s) for s in State],
+                           default='findable',
+                           help="create the DOI with this state")
+create_parser.add_argument('--url', help="URL of the landing page")
+create_parser.add_argument('--metadata',
                            help="XML file with DOI metadata",
                            metavar="metadata.xml",
                            type=Path)
+create_parser.add_argument('doi', help="the DOI to create")
 create_parser.set_defaults(func=create_doi)
 
 
@@ -74,13 +84,18 @@ def update_doi(args):
         metadata.doi = doi.doi
         doi.metadata = metadata
         need_update = True
+    if args.state:
+        need_update = True
     if need_update:
         log.info("Update %s", doi.doi)
-        doi.update(config)
+        doi.update(config, state=args.state)
     else:
         log.info("Nothing to do")
 
 update_parser = subparsers.add_parser('update', help="Update a DOI")
+update_parser.add_argument('--state',
+                           type=State, choices=[str(s) for s in State],
+                           help="change the state of the DOI")
 update_parser.add_argument('--url', help="URL of the landing page")
 update_parser.add_argument('--metadata',
                            help="XML file with DOI metadata",